chore: logic is shifted to plugin
